+++ common/src/main/java/dk/magnusjensen/bedfallback/CommonClass.java
@@ -13,11 +13,12 @@ package dk.magnusjensen.bedfallback;
 
 import dk.magnusjensen.bedfallback.data.BedFallbackSavedData;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.GlobalPos;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.tags.BlockTags;
+import net.minecraft.world.attribute.BedRule;
+import net.minecraft.world.attribute.EnvironmentAttributes;
 import net.minecraft.world.level.block.BedBlock;
-import net.minecraft.world.level.block.entity.BedBlockEntity;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.properties.BedPart;
 
 public class CommonClass {
@@ -28,38 +29,40 @@ public class CommonClass {
     }
 
     public static void handlePlayerSetSpawn(ServerPlayer player, BlockPos bedPos) {
-        // Check that we are in the overworld to not cause weirdness but rather just not support other dimensions setting bed spawns
-        if (player.level().dimension() != player.level().getServer().overworld().dimension()) {
-            Constants.LOG.info("Skipped setting bed spawn for player {} in dimension {}", player.getName().getString(), player.level().dimension().identifier());
-            return;
-        }
-
-        var data = BedFallbackSavedData.getData(player.level());
-
         var blockState = player.level().getBlockState(bedPos);
         if (!blockState.is(BlockTags.BEDS)) {
             return;
         }
 
+        // Whether a bed sets spawn is an environment attribute, so it answers for modded dimensions as well, and it
+        // varies by position rather than by dimension. The Nether and the End say no through the rule that blows beds up.
+        BedRule bedRule = player.level().environmentAttributes().getValue(EnvironmentAttributes.BED_RULE, bedPos);
+        if (!bedRule.canSetSpawn(player.level())) {
+            Constants.LOG.info("Skipped setting bed spawn for player {} in dimension {}, beds do not set spawn there", player.getName().getString(), player.level().dimension().identifier());
+            return;
+        }
+
+        var data = BedFallbackSavedData.getData(player.level().getServer());
+
         // We don't need to check block positions, as the BedBlock class handles always calling code with it's HEAD part.
-        data.addBedSpawnPosition(player.getUUID(), bedPos);
+        data.addBedSpawnPosition(player.getUUID(), GlobalPos.of(player.level().dimension(), bedPos));
     }
 
-    public static void handleBlockBroken(ServerPlayer player, BlockPos brokenPos, BlockEntity blockEntity) {
-        if (!(blockEntity instanceof BedBlockEntity)) {
-            // We can rely on that beds have block entities, so if it's not a bed block entity, we can just return.
+    public static void handleBlockBroken(ServerPlayer player, BlockPos brokenPos) {
+        // Beds lost their block entity in 26.2, so the block state is the only thing left to recognise one by.
+        var state = player.level().getBlockState(brokenPos);
+        if (!state.is(BlockTags.BEDS)) {
             return;
         }
 
         // Always get the head part to ensure consistency with other code pieces
-        var state = player.level().getBlockState(brokenPos);
         BedPart bedPart = state.getValue(BedBlock.PART);
         if (bedPart == BedPart.FOOT) {
             brokenPos = brokenPos.relative(BedBlock.getConnectedDirection(state));
         }
 
-        var data = BedFallbackSavedData.getData(player.level());
+        var data = BedFallbackSavedData.getData(player.level().getServer());
 
-        data.removeBedSpawnPosition(brokenPos);
+        data.removeBedSpawnPosition(GlobalPos.of(player.level().dimension(), brokenPos));
     }
 }